# src/VeyonCampus.Telemetry.Server/openapi/deployment-packages.yaml
openapi: 3.1.0
info:
  title: Veyon Campus Deployment Package API
  version: 1.0.0
  description: >-
    Teacher-published, server-validated Veyon Campus v3 configuration packages.
    Publishing is anonymous and requires campus name, teacher name, and the
    teacher phone's last four digits. The raw phone suffix is not stored; a
    keyed fingerprint is used to verify student downloads. The teacher name is
    retained privately for operational attribution. Mutations and downloads
    use Cache-Control: no-store.
servers:
  - url: https://{gatewayHost}
    description: CloudBase HTTP Gateway routed to the veyon-api HTTP Cloud Function
    variables:
      gatewayHost:
        default: veyon-control-d3gs8hmuyd09c00a7-1348081197.ap-shanghai.app.tcloudbase.com
tags:
  - name: Catalog
  - name: Publishing
security: []
paths:
  /v1/deployment-packages:
    get:
      tags: [Catalog]
      operationId: searchDeploymentPackages
      summary: Search published deployment packages
      parameters:
        - name: query
          in: query
          required: false
          schema: { type: string, maxLength: 100 }
          description: Case-insensitive substring search over campus name and computer prefix.
        - name: campusId
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 49, default: 20 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 10000, default: 0 }
      responses:
        '200':
          description: Published catalog page; never includes private storage keys.
          content:
            application/json:
              schema:
                type: object
                required: [items, limit, offset, hasMore]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PackageCatalogItem' }
                  limit: { type: integer }
                  offset: { type: integer }
                  hasMore: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [Publishing]
      operationId: publishDeploymentPackage
      summary: Publish a validated package with campus and teacher details
      description: >-
        No teacher account or login is required. Provide the campus name, teacher
        name, and teacher phone's last four digits. The phone suffix is used by
        students to download this package; the raw digits are never stored. The
        teacher name is retained privately and is not returned in the student
        catalog. Package content is restricted to public configuration files
        and validated again by the server. The ZIP is limited to 64 KiB, each
        folder file to 16 KiB, and the request to 128 KiB.
        Submit either one ZIP in the archive field, or up to five files in the
        files field. Folder uploads may include one shared top-level folder.
        The server validates and re-creates a canonical ZIP; client filenames
        and object keys are not used as storage names. The CloudBase HTTP gateway
        route is limited to 100 total requests per second.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [campusName, publisherName, teacherPhoneLast4]
              properties:
                campusName: { type: string, minLength: 1, maxLength: 100 }
                publisherName:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: Teacher name retained privately for operational attribution.
                teacherPhoneLast4:
                  type: string
                  pattern: '^[0-9]{4}$'
                  description: Used by students to verify and download this package; raw digits are not stored.
                archive:
                  type: string
                  format: binary
                  description: One schema v3 ZIP, at most 64 KiB.
                files:
                  type: array
                  maxItems: 5
                  items: { type: string, format: binary }
                  description: Flat package folder files; each at most 16 KiB, with at most 64 KiB in total.
            encoding:
              files:
                style: form
                explode: true
      responses:
        '201':
          description: Package object and metadata were published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishedPackage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { description: Request body exceeds 128 KiB. }
        '429': { description: The CloudBase gateway route limit of 100 total requests per second was reached. }
        '502': { $ref: '#/components/responses/CloudBaseRejected' }
        '503': { $ref: '#/components/responses/Unavailable' }

  /v1/deployment-packages/{packageId}/download:
    post:
      tags: [Catalog]
      operationId: downloadDeploymentPackage
      summary: Download a published package after teacher-phone-suffix verification
      parameters:
        - $ref: '#/components/parameters/PackageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teacherPhoneLast4]
              properties:
                teacherPhoneLast4: { type: string, pattern: '^[0-9]{4}$' }
      responses:
        '200':
          description: Canonical ZIP; content is checked against database size and SHA-256 before delivery.
          headers:
            Content-Disposition:
              schema: { type: string }
              description: Attachment filename generated from packageId.
          content:
            application/zip:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { description: Ten failed attempts for one package from the same source within 15 minutes trigger a 15-minute process-local block. A successful code resets the failure window; horizontal instances do not share this in-memory limit. Retry after the Retry-After interval. }
        '502': { $ref: '#/components/responses/ArtifactUnavailable' }
        '503': { $ref: '#/components/responses/Unavailable' }

  /v1/deployment-packages/{packageId}/withdraw:
    post:
      tags: [Publishing]
      operationId: withdrawDeploymentPackage
      summary: Withdraw a published package
      description: Previously authenticated publishers or owner/admin accounts can withdraw older records; anonymously published packages require owner/admin withdrawal.
      security:
        - CloudBaseAccessToken: []
      parameters:
        - $ref: '#/components/parameters/PackageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 500 }
      responses:
        '204': { description: Package withdrawn. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/CloudBaseRejected' }
        '503': { $ref: '#/components/responses/Unavailable' }

components:
  securitySchemes:
    CloudBaseAccessToken:
      type: http
      scheme: bearer
      bearerFormat: CloudBase Auth access_token
  parameters:
    PackageId:
      name: packageId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    PackageCatalogItem:
      type: object
      required:
        - packageId
        - campusId
        - displayName
        - campusName
        - computerPrefix
        - schemaVersion
        - targetOs
        - architecture
        - fileName
        - sizeBytes
        - sha256
        - downloadCount
        - publishedAt
        - requiresPhoneLast4
      properties:
        packageId: { type: string, format: uuid }
        campusId: { type: integer, format: int64, nullable: true }
        displayName: { type: string }
        campusName: { type: string }
        computerPrefix: { type: string, maxLength: 12 }
        schemaVersion: { type: integer, const: 3 }
        targetOs: { type: string, const: windows }
        architecture: { type: string, const: x64 }
        fileName: { type: string, pattern: '^veyon-campus-config-v3-[a-f0-9]{32}\\.zip$' }
        sizeBytes: { type: integer, minimum: 1, maximum: 65536 }
        sha256: { type: string, pattern: '^[A-F0-9]{64}$' }
        downloadCount: { type: integer, format: int64, minimum: 0 }
        publishedAt: { type: string, format: date-time }
        requiresPhoneLast4: { type: boolean, const: true }
    PublishedPackage:
      type: object
      required: [packageId, campusId, campusName, publisherName, computerPrefix, schemaVersion, targetOs, architecture, fileName, sizeBytes, sha256]
      properties:
        packageId: { type: string, format: uuid }
        campusId: { type: integer, format: int64, minimum: 1, nullable: true }
        campusName: { type: string }
        publisherName: { type: string, description: Returned only to the caller that just published this package. }
        computerPrefix: { type: string }
        schemaVersion: { type: integer, const: 3 }
        targetOs: { type: string, const: windows }
        architecture: { type: string, const: x64 }
        fileName: { type: string }
        sizeBytes: { type: integer, minimum: 1, maximum: 65536 }
        sha256: { type: string }
    Problem:
      type: object
      properties:
        type: { type: string, format: uri-reference }
        error: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
  responses:
    BadRequest:
      description: Invalid query, package contents, or request fields.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthenticated:
      description: Missing, invalid, or expired CloudBase Auth access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Teacher phone suffix verification failed, or the operation is not authorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Active campus or published package was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: The packageId has already been published.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ArtifactUnavailable:
      description: Stored object is missing or does not match its recorded hash and size.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    CloudBaseRejected:
      description: CloudBase rejected the publication or withdrawal database operation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unavailable:
      description: CloudBase Auth, PostgreSQL, or PG Storage is temporarily unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }